Feat: generate default config
