+++ package.yaml
@@ -49,11 +49,13 @@ dependencies:
 - rio >= 0.1.1.0
 - aeson
 - extensible >= 0.6
+- file-embed
 - fallible
 - language-docker >= 8.0
 - mix
 - mix-json-logger
 - req >= 2.0
+- th-lift-instances
 - yaml
 
 library:
